+++ sample/apps/wgpu_deferred_rendering/utils.h
@@ -27,7 +27,7 @@ struct GenerateNormalsResult
 {
     std::vector<glm::vec3> positions;
     std::vector<glm::vec3> normals;
-    std::vector<glm::ivec3> triangles;
+    std::vector<glm::i16vec3> triangles;
 };
 
 // --------------------------------------------------------
